Re-factor chat.jspx include and make it one page for both desktop and mobile.
Modified desktop chatview to hold 5 entries instead of 7 and decreased inputText maxlength from 150 to 70, re-styled so that the chat box behaves similar in desktop as it does on mobile.  It shouldn't shift now when different length entries are input and if someone inputs one long string, it simply bleeds to the right.